Refactor dropdowns components
