+++ app/scripts/controllers/network/util.js
@@ -27,12 +27,15 @@ const {
   DROPDOWN_ROPSTEN_DISPLAY_NAME,
   DROPDOWN_RINKEBY_DISPLAY_NAME,
   DROPDOWN_KOVAN_DISPLAY_NAME,
+  chainTypes,
 } = require('./enums')
 
+const { TEST, PROD } = chainTypes
 const networks = {}
 
 const POA_OBJ = {
   order: 1,
+  chainType: PROD,
   providerName: POA,
   networkID: POA_CODE,
   displayName: POA_DISPLAY_NAME,
@@ -43,6 +46,7 @@ networks[POA] = POA_OBJ
 
 const DAI_OBJ = {
   order: 2,
+  chainType: PROD,
   providerName: DAI,
   networkID: DAI_CODE,
   displayName: DAI_DISPLAY_NAME,
@@ -53,6 +57,7 @@ networks[DAI] = DAI_OBJ
 
 const POA_SOKOL_OBJ = {
   order: 3,
+  chainType: TEST,
   providerName: POA_SOKOL,
   networkID: POA_SOKOL_CODE,
   displayName: POA_SOKOL_DISPLAY_NAME,
@@ -63,6 +68,7 @@ networks[POA_SOKOL] = POA_SOKOL_OBJ
 
 const MAINNET_OBJ = {
   order: 4,
+  chainType: PROD,
   providerName: MAINNET,
   networkID: MAINNET_CODE,
   displayName: MAINNET_DISPLAY_NAME,
@@ -73,6 +79,7 @@ networks[MAINNET] = MAINNET_OBJ
 
 const ROPSTEN_OBJ = {
   order: 5,
+  chainType: TEST,
   providerName: ROPSTEN,
   networkID: ROPSTEN_CODE,
   displayName: ROPSTEN_DISPLAY_NAME,
@@ -83,6 +90,7 @@ networks[ROPSTEN] = ROPSTEN_OBJ
 
 const KOVAN_OBJ = {
   order: 6,
+  chainType: TEST,
   providerName: KOVAN,
   networkID: KOVAN_CODE,
   displayName: KOVAN_DISPLAY_NAME,
@@ -93,6 +101,7 @@ networks[KOVAN] = KOVAN_OBJ
 
 const RINKEBY_OBJ = {
   order: 7,
+  chainType: TEST,
   providerName: RINKEBY,
   networkID: RINKEBY_CODE,
   displayName: RINKEBY_DISPLAY_NAME,
