+++ old-ui/app/account-detail.js
@@ -13,7 +13,7 @@ const ethUtil = require('ethereumjs-util')
 const EditableLabel = require('./components/editable-label')
 const TabBar = require('./components/tab-bar')
 const TokenList = require('./components/token-list')
-const AccountDropdowns = require('./components/account-dropdowns').AccountDropdowns
+const AccountDropdowns = require('./components/account-dropdowns/index').AccountDropdowns
 const CopyButton = require('./components/copyButton')
 const ToastComponent = require('./components/toast')
 import { getMetaMaskAccounts } from '../../ui/app/selectors'
